production/requirements.txt
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0
tqdm>=4.65.0
scikit-learn>=1.3.0
scipy>=1.10.0
matplotlib>=3.8.0
seaborn>=0.13.0
pytest>=7.4.0
pytest-cov>=4.1.0
